Displaying Linked List
Now displaying the linked list, with an initial head of the linked list being an example receipt. Created buttons to utilise the add and remove functions, which both work but once items are added or removed there is no change in the display. Also got rid of the FileView as it was not needed.